+++ test/samplers.jl
@@ -9,7 +9,7 @@ using TaijaBase.Samplers:
     ConditionalSampler,
     UnconditionalSampler,
     AbstractSampler,
-    PCD
+    PMC
 
 @testset "Samplers" begin
 
@@ -49,7 +49,7 @@ using TaijaBase.Samplers:
         end
     end
 
-    @testset "Persistent Contrastive Divergence (PCD)" begin
+    @testset "Persistent Markov Chains (PMC)" begin
 
         # Train a simple neural network on the data (classification)
         Xtrain = MLJBase.matrix(X) |> permutedims
@@ -69,18 +69,18 @@ using TaijaBase.Samplers:
             println("Accuracy: ", mean(Flux.onecold(nn(Xtrain)) .== Flux.onecold(ytrain)))
         end
 
-        # PCD
+        # PMC
         bs = 10
         ntrans = 100
         niter = 20
         # Conditionally sample from first class:
         smpler =
             ConditionalSampler(𝒟x, 𝒟y, input_size = size(Xmat)[1:end-1], batch_size = bs)
-        x1 = PCD(smpler, nn, ImproperSGLD(); ntransitions = ntrans, niter = niter, y = 1)
+        x1 = PMC(smpler, nn, ImproperSGLD(); ntransitions = ntrans, niter = niter, y = 1)
         # Conditionally sample from second class:
         smpler =
             ConditionalSampler(𝒟x, 𝒟y, input_size = size(Xmat)[1:end-1], batch_size = bs)
-        x2 = PCD(smpler, nn, ImproperSGLD(); ntransitions = ntrans, niter = niter, y = 2)
+        x2 = PMC(smpler, nn, ImproperSGLD(); ntransitions = ntrans, niter = niter, y = 2)
 
         # using Plots
         # plt = scatter(Xtrain[1, :], Xtrain[2, :], color=Int.(y.refs), group=Int.(y.refs), label=["X|y=0" "X|y=1"], alpha=0.1)